+++ src/pages/CharacterDetail/interfaces/CharacterDetail.ts
@@ -1,0 +1,12 @@
+import type {
+  Character,
+  Origin,
+  Location,
+} from "../../../shared/interfaces/Character";
+
+export interface CharacterDetail {
+  character: Character & {
+    origin: Origin;
+    location: Location;
+  };
+}