+++ src/pages/characters/components/Actions.tsx
@@ -1,10 +1,12 @@
-export const Actions = ({ action }: { action: string }) => {
+import { Link } from "react-router";
+
+export const Actions = ({ action, id }: { action: string; id: string }) => {
   return (
-    <button
-      type="button"
+    <Link
+      to={`/characters/${id}`}
       className="bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded text-sm"
     >
       {action}
-    </button>
+    </Link>
   );
 };