+++ src/shared/components/Status/Status.tsx
@@ -1,11 +1,9 @@
-import type { Character } from "../../../../shared/interfaces/Character";
+import type { Character } from "../../interfaces/Character";
 import { StatusFactory } from "./StatusFactory";
 
 export const Status = ({ status }: { status: Character["status"] }) => {
   return (
-    <div
-      className="text-gray-300 text-sm font-normal leading-relaxed text-center"
-    >
+    <div className="text-gray-300 text-sm font-normal leading-relaxed text-center">
       <StatusFactory status={status} />
     </div>
   );